Move two errors to base
