+++ src/certs/capabilities/mod.rs
@@ -13,7 +13,7 @@ use der::asn1::SetOfVec;
 
 use x509_cert::attr::{Attribute, Attributes};
 
-use crate::errors::composite::InvalidInput;
+use crate::errors::base::InvalidInput;
 use crate::{Constrained, ConstraintError};
 
 /// Object Identifier for the KeyUsage::DigitalSignature variant.
