+++ src/errors/base.rs
@@ -16,3 +16,27 @@ pub enum ConstraintError {
         reason: Option<String>,
     },
 }
+
+/// Represents errors for invalid input in IdCsr or IdCert generation.
+#[derive(Error, Debug, PartialEq, Clone)]
+pub enum InvalidInput {
+    #[error("The der library has reported the following error with the input")]
+    DerError(der::Error),
+    #[error("subject_session_id MUST NOT exceed length limit of 32 characters")]
+    SessionIdTooLong,
+    #[error(
+        "Cannot perform conversion, as input variant can not be converted to output. {reason:}"
+    )]
+    IncompatibleVariantForConversion { reason: String },
+}
+
+#[derive(Error, Debug, PartialEq, Clone)]
+// TODO: Replace usages of InvalidInput::IncompatibleVariantForConversion with this Enum
+pub enum UnsuccessfulConversion {
+    #[error(
+        "Cannot perform conversion, as input variant can not be converted to output. {reason:}"
+    )]
+    IncompatibleVariant { reason: String },
+    #[error("Conversion failed due to invalid input")]
+    InvalidInput(String),
+}
